moved model functions into one model as requested
