+++ src/models/PostsModel.php
@@ -24,4 +24,13 @@ class PostsModel{
         }
         return false;
     }
+
+    public function addPost($title, $content, $userID): bool
+    {
+        $query = $this->db->prepare("INSERT INTO `posts` (`title`, `content`, `user_id`, `date_time`) VALUES (:title, :content, :user_id, NOW());");
+        if ($query->execute([':title' => $title, ':content' => $content, ':user_id' => $userID])) {
+            return true;
+        }
+        return false;
+    }
 }
